chore: rename platform -> camunda-platform
