+++ src/camunda-platform/extractors/extractEventDefinitionVariables.js
