+++ src/camunda-platform/extractors/extractOutMappings.js
