+++ src/camunda-platform/extractors/extractOutputParameters.js
