+++ src/camunda-platform/extractors/extractResultVariables.js
