+++ src/camunda-platform/util/ElementsUtil.js
