+++ src/camunda-platform/util/EventDefinitionsUtil.js
