+++ src/camunda-platform/util/ExtensionElementsUtil.js
