+++ src/camunda-platform/util/ProcessVariablesUtil.js
