+++ test/camunda-platform/fixtures/call-activity-local.bpmn
