+++ test/camunda-platform/fixtures/error-code-variable.bpmn
