+++ test/camunda-platform/fixtures/error-message-variable.bpmn
