+++ test/camunda-platform/fixtures/result-variable.bpmn
