+++ test/camunda-platform/fixtures/sub-process-output-mapping.bpmn
