+++ test/camunda-platform/fixtures/sub-process-own-scope.bpmn
